daisy/requirements.txt
pyserial==3.5
schedule==1.1.0